Rigidbody Physics Controller Update
Updates have been made to the basic physics controller to allow for more natural movement by changing the movement type from a force to a position translation.

Note there is a small bug in where when the player collides with something, it will start to slowly rotate.

I am working on a fix for this using angular drag, but will research other functions.